update project dependencies
Refresh Cargo, npm, GitHub Actions, Rust, Rig, and Nix dependency inputs. Adapt
tool result correlation, SQLite integer bindings, and React types to their
updated APIs while preserving existing runtime behavior.

React 19 changes the JSX namespace and nullable ref types. Rig uses typed tool
call identifiers and requires tool names on results. Rusqlite 0.40 requires
SQLite-compatible signed integer bindings for platform-sized values.

+++ src/database/schema.rs
@@ -188,7 +188,7 @@ pub(super) fn migrate(connection: &Connection) -> Result<(), DatabaseError> {
         transaction.execute_batch(sql)?;
         transaction.execute(
             "INSERT INTO schema_migrations(version, applied_at) VALUES (?1, ?2)",
-            params![index + 1, timestamp(Utc::now())],
+            params![(index + 1) as i64, timestamp(Utc::now())],
         )?;
         transaction.commit()?;
     }
